rename subcommand bench to test; add supports for differential test (#72)

+++ cmd/explain.go
@@ -30,7 +30,7 @@ func explainCommand() *cli.Command {
 		Aliases: []string{"e"},
 		Usage:   "Explain analyze a query",
 		Flags: []cli.Flag{
-			&cli.Int64Flag{
+			&cli.Uint64Flag{
 				Name:        "plan",
 				Aliases:     []string{"p"},
 				Usage:       "use plan by `ID`",
@@ -56,7 +56,7 @@ func explainCommand() *cli.Command {
 				return err
 			}
 
-			plan, err := horoscope.Plan(query, hints, options.Query.PlanID)
+			plan, err := horoscope.Plan(query, hints, int64(options.Query.PlanID))
 			if err != nil {
 				return err
 			}
